+++ CHANGELOG.md
@@ -5,6 +5,10 @@
   ## **WORK IN PROGRESS**
   * (theimo1221) Update packages
 -->
+## **WORK IN PROGRESS**
+
+* (nockiro) Add Osram dimmer device
+* 
 ## 1.1.42 (2022-10-15)
 
 * (theimo1221) Update packages